Import changes from GUWA project. (#266)
- Import changes from GUWA project.
- Derive bp_gaussian beam pattern from bp_solid.
- Add -fPIC to USMLUse.cmake file
- Change platform IDs from int to uint64_t to better support JSE/GRID ids.
- Ignore -Wmaybe-uninitialized from boost::geomentry in eigenverb_collection.h
- Add ability to read USML_DATA_DIR from environment variables in ocean_utils::make_basic() function.
- Add is_acoustic_target flag to platform_model so that it can be skipped in wavefront generation.
- Fix index calculation in rvbts_collection.cc file.
- Add ability to read time_series data from rvbts_collection.
- Add treverb parameter to rvbts_generator as replacement for fsample.
- Add ability to reset_src_beams() in sensor_model.
- Get rid of fsample and fcenter in sensor_model.
- Fix the way that wavefront generators are computed.
- Fix the way that thead_controller is reset.
- Minor clang-tidy fixes.
- Resolves #263 "git-lfs" undocumented requirement.
- Resolves #269 reverb_analytic_test fails to compile
- Resolves #265 Invalid array index in data_grid_svp.h

+++ eigenrays/eigenray_collection.h
@@ -49,8 +49,8 @@ class USML_DECLSPEC eigenray_collection : public eigenray_listener {
      */
     eigenray_collection(const seq_vector::csptr &frequencies,
                         const wposition1 &source_pos,
-                        const wposition &target_pos, int sourceID = 0,
-                        const matrix<int> &targetIDs = matrix<int>(),
+                        const wposition &target_pos, uint64_t sourceID = 0,
+                        const matrix<uint64_t> &targetIDs = matrix<uint64_t>(),
                         bool coherent = true);
 
     /**
@@ -82,7 +82,7 @@ class USML_DECLSPEC eigenray_collection : public eigenray_listener {
     }
 
     /// Platform ID number for this source. Set to zero if unknown.
-    int sourceID() const { return _sourceID; }
+    uint64_t sourceID() const { return _sourceID; }
 
     /**
      * Platform ID number for one target. Set to zero if unknown.
@@ -91,7 +91,7 @@ class USML_DECLSPEC eigenray_collection : public eigenray_listener {
      * @param   t2  		Column number of target.
      * @return  Value to find target in platform_manager.
      */
-    int targetID(size_t t1 = 0, size_t t2 = 0) const {
+    uint64_t targetID(size_t t1 = 0, size_t t2 = 0) const {
         if (size1() == 0 && size2() == 0) {
             return 0;
         }
@@ -123,7 +123,7 @@ class USML_DECLSPEC eigenray_collection : public eigenray_listener {
      * @param   targetID	  Platform ID number for this target.
      * @return  List of acoustic paths between a source and target.
      */
-    eigenray_list find_eigenrays(int targetID = 0) const;
+    eigenray_list find_eigenrays(uint64_t targetID = 0) const;
 
     /**
      * Find fastest eigenray for a single target in the grid.
@@ -131,7 +131,7 @@ class USML_DECLSPEC eigenray_collection : public eigenray_listener {
      * @param   targetID	  Platform ID number for this target.
      * @return  Time of arrival of the fastest eigenray, zero if not found.
      */
-    double find_initial_time(int targetID = 0) const;
+    double find_initial_time(uint64_t targetID = 0) const;
 
     /**
      * Propagation loss for a single target summed over eigenrays.
@@ -315,10 +315,10 @@ class USML_DECLSPEC eigenray_collection : public eigenray_listener {
 
    private:
     /// Value to find source in platform_manager. Set to zero if unknown.
-    const int _sourceID;
+    const uint64_t _sourceID;
 
     /// Value to find targets in platform_manager. Set to zero if unknown.
-    matrix<int> _targetIDs;
+    matrix<uint64_t> _targetIDs;
 
     /**
      * Location of the wavefront source in spherical earth coordinates.
