+++ wavegen/wavefront_generator.cc
@@ -26,7 +26,7 @@ using namespace usml::waveq3d;
  */
 wavefront_generator::wavefront_generator(
     sensor_model* source, const wposition& target_positions,
-    const matrix<int>& targetIDs, const seq_vector::csptr& frequencies,
+    const matrix<uint64_t>& targetIDs, const seq_vector::csptr& frequencies,
     const seq_vector::csptr& de_fan, const seq_vector::csptr& az_fan,
     double time_step, double time_maximum, double intensity_threshold,
     int max_bottom, int max_surface)
